Refactor: 엔드포인트 상수화
